Module kfs_libuser::error

Error handling

Errors in Horizon/NX follow a specific format. They are encoded on a 32-bit integer, where the bottom 9 bits represent the module, while the top bits represent the "description". A Module is usually a sysmodule, with a few additional modules for specific libraries (Module 168 is userland crash, for instance). See the switchbrew Error Codes page for more information.

Such errors are nice, but have one small problem: they're missing backtraces. In libuser, we opted to use a failure enum. We have a different enum for every module, listing all of their error descriptions, and a big enum over all those module errors. This fine-grained approach makes error handling code nicer. For instance, writing a function returning an Error:

fn ret_err() -> Result<(), Error> {
   // Will automatically be converted to Error, the backtrace filled
   let _ = Err(KernelError::InvalidPort)?;
   let _ = Err(SmError::PermissionDenied)?;
   Ok(())
}

Matching on an error is similarly convenient:

match err {
   Ok(_) => (),
   Err(Error::Kernel(KernelError::InvalidPort, _)) => (),
   _ => ()
}

Structs

KernelError
Kernel syscall error codes.

LibuserError
Internal libuser errors.

Module

SmError
Service Manager errors.

Enums

Error
The global error type. Every error defined here can be downcasted to this type. A Backtrace will be created when casting an error to this type.

Constants

_DERIVE_failure_Fail_FOR_Error