Module kfs_libuser::ipc::server

IPC Server primitives

The creation of an IPC server requires a WaitableManager and a PortHandler. The WaitableManager will manage the event loop: it will wait for a request to arrive on one of the waiters (or for any other event to happen), and call that waiter's handle_signal function.

A PortHandler is a type of Waiter which listens for incoming connections on a port, creates a new Object from it, wrap it in a SessionWrapper (a kind of waiter), and adds it to the WaitableManager's wait list.

When a request comes to the Session, the SessionWrapper's handle_signaled will call the dispatch function of its underlying object.

Here's a very simple example server:

struct IExample;
object! {
    impl IExample {
        #[cmdid(0)]
        fn hello(&mut self, ) -> Result<([u8; 5]), Error> {
             Ok(b"hello")
        }
    }
}

fn main() {
     let man = WaitableManager::new();
     let handler = Box::new(PortHandler::<IExample>::new("hello\0").unwrap());
     man.add_waitable(handler as Box<dyn IWaitable>);
     man.run()
}

Structs

Align16

PortHandler
A wrapper around a Server Port that implements the IWaitable trait. Waits for connection requests, and creates a new SessionWrapper around the incoming connections, which gets registered on the WaitableManager.

SessionWrapper
A wrapper around an Object backed by an IPC Session that implements the IWaitable trait.

WaitableManager
The event loop manager. Waits on the waitable objects added to it.

Traits

IWaitable
A handle to a waitable object.

Object
An IPC object.

Functions

encode_bytes