Struct kfs_libuser::ipc::IPCBuffer

pub struct IPCBuffer<'a> {
    addr: u64,
    size: u64,
    ty: IPCBufferType,
    phantom: PhantomData<&'a ()>,
}

An IPC Buffer represents a section of memory to send to the other side of the pipe. It is usually used for sending big chunks of data that would not send in the comparatively small argument area (which is usually around 200 bytes).

There exists 5 types of IPC Buffers: Send(A), Receive(B), SendReceive(W), Pointer(X) and ReceiveList(C).

Send/Receive/SendReceive buffers work by remapping the memory from the sender's process into the receiver's process. This means that they need to have a page-aligned address and size.

In contrast, Pointer/ReceiveList buffers work by memcpying the sender's Pointer buffer into the receiver's ReceiveList buffer. This allows greater flexibility on the address and size. In general, those are prefered.

Fields
addr: u64

Address to the value

size: u64

Size of the value

ty: IPCBufferType

Buffer type

phantom: PhantomData<&'a ()>

Tie the buffer's lifetime to the value's ! This is very very very important, for the safety of this interface. It ensures that, as long as this IPCBuffer exist, the value it references cannot be dropped.

Methods

impl<'a> IPCBuffer<'a>

fn in_pointer<T>(data: &mut T, has_u16_size: bool) -> IPCBuffer

Creates a Type-C IPCBuffer from the given reference.

If has_u16_size is true, the size of the pointer will be written after the raw data. This is only used when sending client-sized arrays.

fn out_pointer<T>(data: &T, counter: u8) -> IPCBuffer

Creates a Type-X IPCBuffer from the given reference.

The counter defines which type-C buffer this should be copied into.

fn buftype(&self) -> IPCBufferType

Gets the IPCBufferType of this buffer. The buffer type determines how the buffer is passed to the other process and if it's an argument or a return value.