Macro kfs_libuser::object

macro_rules! object {
    (impl $tyname:ident {
        $($fns:tt)*
    }) => { ... };
    (@functions meta=($($meta:meta),*), #[cmdid($expr:expr)] $($tt:tt)*) => { ... };
    (@functions meta=($($meta:meta),*), #[$fmeta:meta] $($tt:tt)*) => { ... };
    (@functions meta=($($meta:meta),*), fn $funcname:ident $(<$($lifetime:lifetime),*>)* (&mut $self:ident, $($args:tt)*) -> Result<($($ret:tt)*), Error> $body:block $($tt:tt)*) => { ... };
    (@functions meta=(), ) => { ... };
    (@rewriteargs ($($meta:meta),*), $funcname:ident, ($($lifetime:lifetime),*), $self:ident, ($($argname:ident : $argty:ty),*), ($($ret:tt)*), $body:block, $name:ident : Handle<copy>, $($rest:tt)*) => { ... };
    (@rewriteargs ($($meta:meta),*), $funcname:ident, ($($lifetime:lifetime),*), $self:ident, ($($argname:ident : $argty:ty),*), ($($ret:tt)*), $body:block, $name:ident : Handle<move>, $($rest:tt)*) => { ... };
    (@rewriteargs ($($meta:meta),*), $funcname:ident, ($($lifetime:lifetime),*), $self:ident, ($($argname:ident : $argty:ty),*), ($($ret:tt)*), $body:block, $name:ident : $ty:ty, $($rest:tt)*) => { ... };
    (@rewriteargs ($($meta:meta),*), $funcname:ident, ($($lifetime:lifetime),*), $self:ident, ($($args:tt)*), ($($ret:tt)*), $body:block, ) => { ... };
    (@dispatch $sel:expr, $manager:expr, cmdid=$cmdid:expr, buf=$buf:expr, fns=($($fns:tt)*), #[cmdid($fcmdid:expr)] $($tt:tt)*) => { ... };
    (@dispatch $sel:expr, $manager:expr, cmdid=$cmdid:expr, buf=$buf:expr, fns=($($fns:tt)*), #[$fmeta:meta] $($tt:tt)*) => { ... };
    (@dispatch $sel:expr, $manager:expr, cmdid=$cmdid:expr, buf=$buf:expr, fns=($($fns:tt)*), fcmdid=$fcmdid:expr, #[$fmeta:meta] $($tt:tt)*) => { ... };
    (@dispatch $sel:expr, $manager:expr, cmdid=$cmdid:expr, buf=$buf:expr, fns=($($fcmdid:expr => $fcmdfn:block),*), fcmdid=$curfcmdid:expr,
       fn $funcname:ident $(<$($lifetime:lifetime),*>)* (&mut self, $($args:tt)*) -> Result<($($ret:tt)*), Error> $body:block $($tt:tt)*) => { ... };
    (@dispatch $sel:expr, $manager:expr, cmdid=$cmdid:expr, buf=$buf:expr, fns=($($fcmdid:expr => $fcmd:block),*), ) => { ... };
    (@genstruct $ident:ident fields=($($args:tt)*), $name:ident: InBuffer<$ty:ty>, $($tt:tt)*) => { ... };
    (@genstruct $ident:ident fields=($($args:tt)*), $name:ident: OutBuffer<$ty:ty>, $($tt:tt)*) => { ... };
    (@genstruct $ident:ident fields=($($args:tt)*), $name:ident: InPointer<$ty:ty>, $($tt:tt)*) => { ... };
    (@genstruct $ident:ident fields=($($args:tt)*), $name:ident: OutPointer<$ty:ty>, $($tt:tt)*) => { ... };
    (@genstruct $ident:ident fields=($($args:tt)*), $name:ident: Handle<move>, $($tt:tt)*) => { ... };
    (@genstruct $ident:ident fields=($($args:tt)*), $name:ident: Handle<copy>, $($tt:tt)*) => { ... };
    (@genstruct $ident:ident fields=($($args:tt)*), $name:ident: Pid, $($tt:tt)*) => { ... };
    (@genstruct $ident:ident fields=($($args:tt)*), $name:ident: &WaitableManager, $($tt:tt)*) => { ... };
    (@genstruct $ident:ident fields=($($iname:ident: $ity:ty),*), $name:ident: $ty:ty, $($tt:tt)*) => { ... };
    (@genstruct $ident:ident fields=($($name:ident: $ty:ty),*), ) => { ... };
    (@callargs $sel:expr, $manager:expr, funcname=$funcname:ident, msgin=$msgin:expr, args=($($arg:expr),*), $name:ident: InBuffer<$ty:ty>, $($tt:tt)*) => { ... };
    (@callargs $sel:expr, $manager:expr, funcname=$funcname:ident, msgin=$msgin:expr, args=($($arg:expr),*), $name:ident: OutBuffer<$ty:ty>, $($tt:tt)*) => { ... };
    (@callargs $sel:expr, $manager:expr, funcname=$funcname:ident, msgin=$msgin:expr, args=($($arg:expr),*), $name:ident: InPointer<$ty:ty>, $($tt:tt)*) => { ... };
    (@callargs $sel:expr, $manager:expr, funcname=$funcname:ident, msgin=$msgin:expr, args=($($arg:expr),*), $name:ident: OutPointer<$ty:ty>, $($tt:tt)*) => { ... };
    (@callargs $sel:expr, $manager:expr, funcname=$funcname:ident, msgin=$msgin:expr, args=($($arg:expr),*), $name:ident: Handle<move>, $($tt:tt)*) => { ... };
    (@callargs $sel:expr, $manager:expr, funcname=$funcname:ident, msgin=$msgin:expr, args=($($arg:expr),*), $name:ident: Handle<copy>, $($tt:tt)*) => { ... };
    (@callargs $sel:expr, $manager:expr, funcname=$funcname:ident, msgin=$msgin:expr, args=($($arg:expr),*), $name:ident: Pid, $($tt:tt)*) => { ... };
    (@callargs $sel:expr, $manager:expr, funcname=$funcname:ident, msgin=$msgin:expr, args=($($arg:expr),*), $name:ident: &WaitableManager, $($tt:tt)*) => { ... };
    (@callargs $sel:expr, $manager:expr, funcname=$funcname:ident, msgin=$msgin:expr, args=($($arg:expr),*), $name:ident: $ty:ty, $($tt:tt)*) => { ... };
    (@callargs $sel:expr, $manager:expr, funcname=$funcname:ident, msgin=$msgin:expr, args=($($arg:expr),*), ) => { ... };
    (@genret fields=($($args:tt)*), retfields=($($rname:ident => $rfn:expr),*), $msgout:expr, $ret:expr, InBuffer<$ty:ty>, $($tt:tt)*) => { ... };
    (@genret fields=($($args:tt)*), retfields=($($rname:ident => $rfn:expr),*), $msgout:expr, $ret:expr, OutBuffer<$ty:ty>, $($tt:tt)*) => { ... };
    (@genret fields=($($args:tt)*), retfields=($($rname:ident => $rfn:expr),*), $msgout:expr, $ret:expr, InPointer<$ty:ty>, $($tt:tt)*) => { ... };
    (@genret fields=($($args:tt)*), retfields=($($rname:ident => $rfn:expr),*), $msgout:expr, $ret:expr, OutPointer<$ty:ty>, $($tt:tt)*) => { ... };
    (@genret fields=($($args:tt)*), retfields=($($rname:ident => $rfn:expr),*), $msgout:expr, $ret:expr, Handle, $($tt:tt)*) => { ... };
    (@genret fields=($($args:tt)*), retfields=($($rname:ident => $rfn:expr),*), $msgout:expr, $ret:expr, HandleRef<$lifetime:lifetime>, $($tt:tt)*) => { ... };
    (@genret fields=($($args:tt)*), retfields=($($rname:ident => $rfn:expr),*), $msgout:expr, $ret:expr, Pid, $($tt:tt)*) => { ... };
    (@genret fields=($($iname:ident: $ity:ty),*), retfields=($($rname:ident => $rfn:expr),*), $msgout:expr, $ret:expr, $ty:ty, $($tt:tt)*) => { ... };
    (@genret fields=($($name:ident: $ty:ty),*), retfields=($($rname:ident => $rfn:expr),*), $msgout:expr, $ret:expr, ) => { ... };
    (@bufcount ) => { ... };
    (@bufcount InBuffer<$ty:ty>, $($tt:tt)*) => { ... };
    (@bufcount OutBuffer<$ty:ty>, $($tt:tt)*) => { ... };
    (@bufcount InPointer<$ty:ty>, $($tt:tt)*) => { ... };
    (@bufcount OutPointer<$ty:ty>, $($tt:tt)*) => { ... };
    (@bufcount $ty:ty, $($tt:tt)*) => { ... };
    (@bufcount $name:ident: $($tt:tt)*) => { ... };
    (@copycount ) => { ... };
    (@copycount HandleRef, $($tt:tt)*) => { ... };
    (@copycount Handle<copy>, $($tt:tt)*) => { ... };
    (@copycount $ty:ty, $($tt:tt)*) => { ... };
    (@copycount $name:ident: $($tt:tt)*) => { ... };
    (@movecount) => { ... };
    (@movecount Handle<move>, $($tt:tt)*) => { ... };
    (@movecount Handle, $($tt:tt)*) => { ... };
    (@movecount $ty:ty, $($tt:tt)*) => { ... };
    (@movecount $name:ident: $($tt:tt)*) => { ... };
}

Auto derive Object.

Takes a service implementation as a parameter, and auto-derives Object, generating the dispatcher logic. The syntax for using it is as under:

use kfs_libuser::error::Error;

struct Service;
object! {
    impl Service {
        #[cmdid(0)]
        fn test(&mut self, val: u32,) -> Result<(u32,), Error> {
            Ok((0,))
        }
    }
}

Allowed argument types:

Pid: the Pid of the remote process, from the Handle Descriptor

Handle: A Handle in the Handle Descriptor's move handle list.

Handle: A Handle in the Handle Descriptor's copy handle list.

InBuffer/OutBuffer/InPointer/OutPointer: Equivalent to A/B/C/X buffers.

&WaitableManager: The manager that signaled us.

T where T: Copy, Clone, repr(C): passed through Raw Data.

Allowed return types:

Pid: The current Pid, from the Handle Descriptor.

Handle: A Handle in the Handle Descriptor's move handle list.

HandleRef: A Handle in the Handle Descriptor's copy handle list.

<T: IObject>: An Object that will be added to the current WaitableManager for processing. Either added to the Handle Descriptor copy handle list, or as a domain object

T where T: Copy, Clone, repr(C): passed through Raw Data.

Important note: the return value has to be a Result<(Type,...), Error>. The function must take &mut self as a first argument. And right now, as a small caveat, the argument list and return tuple list must end with a coma.

The implementation will be left verbatim (except for the cmdid). A separate impl Object for Service block will be generated.

This macro is deeply recursive. If used, you'll likely need to add #![recursion_limit = "1024"] - or more - to your crate root.