Function kfs_libuser::syscalls::create_shared_memory

pub fn create_shared_memory(
    size: usize, 
    myperm: MemoryPermissions, 
    otherperm: MemoryPermissions
) -> Result<SharedMemory, KernelError>

Creates a shared memory handle.

Allocates the given size bytes of physical memory to back the SharedMemory. myperm dictates the memory permissions this handle can be mapped as in the current process, while otherperm dictates the permissions for other processes.

Errors

Errors if size is not page-aligned.