Struct kfs_libuser::terminal::Terminal

pub struct Terminal {
    framebuffer: Window,
    cursor_pos: Pos,
    font: Font<'static>,
    cached_glyphs: HashMap<char, GlyphBitmap>,
    advance_width: usize,
    linespace: usize,
    ascent: usize,
    descent: usize,
}

A struct for logging text to the window. Renders characters from a .ttf font using the font-rs crate

Fields
framebuffer: Window

Rendering target for this terminal.

cursor_pos: Pos

Cursor pos, in pixels. Does not account for bpp. Reprensents the pen position on the baseline.

font: Font<'static>

The font in use for this terminal.

cached_glyphs: HashMap<char, GlyphBitmap>

We cache ascii glyphs to avoid rendering them every time.

advance_width: usize

Expected to be the same for every glyph since it should be a monospaced font.

linespace: usize

The distance between two baselines.

ascent: usize

The maximum ascent in the font.

descent: usize

The maximum descent in the font.

Methods

impl Terminal

pub fn new(size: WindowSize) -> Result<Self, Error>

Creates a new Window of the requested size for terminal usage.

pub fn draw(&mut self) -> Result<(), Error>

Ask the compositor to redraw the window.

fn carriage_return(&mut self)

Move the cursor to the beginning of the current line.

fn line_feed(&mut self)

Move the cursor to the beginning of the next line, scrolling the screen if necessary.

fn advance_pos(&mut self)

Move the cursor to the next position for drawing a character, possibly the next line if we need to wrap.

fn move_pos_back(&mut self)

Move the cursor back to the previous position. If we are already on the first character position on this line, do not move.

fn scroll_screen(&mut self)

scrolls the whole screen by one line. self.pos must be on last baseline.

pub fn clear(&mut self)

Clears the whole screen and reset cursor

fn print_attr(&mut self, string: &str, fg: Color, bg: Color)

Prints a string to the screen with attributes

fn display_glyph_in_box(
glyph: &GlyphBitmap,
framebuffer: &mut Window,
box_width: usize,
box_ascent: usize,
box_descent: usize,
fg: Color,
bg: Color,
pos: Pos
)

Copies a rendered character to the screen, displaying it in a bg colored box

Panics

Panics if pos makes writing the glyph overflow the screen