Struct kfs_libuser::window::Window

pub struct Window {
    buf: MappedSharedMemory,
    handle: IBuffer,
    width: usize,
    height: usize,
    bpp: usize,
}

A managed window.

Fields
buf: MappedSharedMemoryhandle: IBufferwidth: usizeheight: usizebpp: usize
Methods

impl Window

pub fn new(
top: i32,
left: i32,
width: u32,
height: u32
) -> Result<Window, Error>

Creates a window in the vi compositor.

pub fn draw(&mut self) -> Result<(), Error>

Ask the compositor to redraw the window.

pub fn width(&self) -> usize

window width in pixels. Does not account for bpp

pub fn height(&self) -> usize

window height in pixels. Does not account for bpp

pub fn bpp(&self) -> usize

The number of bits that forms a pixel. Used to compute offsets in window memory to corresponding pixel px_offset = px_nbr * bpp

pub fn get_px_offset(&self, x: usize, y: usize) -> usize

Gets the offset in memory of a pixel based on an x and y.

Panics

Panics if y >= self.height() or x >= self.width()

pub fn write_px(&mut self, offset: usize, color: &Color)

Writes a pixel in the window respecting the bgr pattern

Panics

Panics if offset is invalid

pub fn write_px_at(&mut self, x: usize, y: usize, color: &Color)

Writes a pixel in the window respecting the bgr pattern Computes the offset in the window from x and y.

Panics

Panics if coords are invalid

pub fn get_buffer(&mut self) -> &mut [Color]

Gets the underlying framebuffer

pub fn clear(&mut self)

Clears the whole window, making it black.