Struct kfs_libuser::ipc::Message

pub struct Message<'a, RAW, BUFF = [IPCBuffer<'a>; 0], COPY = [u32; 0], MOVE = [u32; 0]> where
    BUFF: Array<Item = IPCBuffer<'a>>,
    COPY: Array<Item = u32>,
    MOVE: Array<Item = u32>,
    RAW: Copy + Default,  {
    ty: u16,
    pid: Option<u64>,
    buffers: ArrayVec<BUFF>,
    copy_handles: ArrayVec<COPY>,
    move_handles: ArrayVec<MOVE>,
    is_request: bool,
    cmdid_error: u32,
    token: Option<u32>,
    raw: RAW,
}

A generic IPC message, representing either an IPC Request or an IPC Response.

In order to ensure performance, the request lives entirely on the stack, no heap allocation is done. However, if we allowed the maximum sizes for everything, this structure would be over-sized, spanning a page. In order to avoid this, we allow the user to set the size of the various parameters they need.

When sending a request that needs to send a COPY handle, the user is expected to create a message specifying the COPY count through the type argument, e.g.

let msg = Message::<_, _, [_; 1], _>::new_request(None, 1);

The ugly syntax, while unfortunate, is a necessary evil until const generics happen.

Fields
ty: u16pid: Option<u64>buffers: ArrayVec<BUFF>copy_handles: ArrayVec<COPY>move_handles: ArrayVec<MOVE>is_request: boolcmdid_error: u32token: Option<u32>raw: RAW
Methods

impl<'a, RAW, BUFF, COPY, MOVE> Message<'a, RAW, BUFF, COPY, MOVE> where
BUFF: Array<Item = IPCBuffer<'a>>,
COPY: Array<Item = u32>,
MOVE: Array<Item = u32>,
RAW: Copy + Default,

pub fn new_request(
token: Option<u32>,
cmdid: u32
) -> Message<'a, RAW, BUFF, COPY, MOVE>

Create a new request for the given cmdid. If a token is passed, the new IPC version will be used. The tokens allow for tracking an IPC request chain. The raw data will contain the default value, and all arrays will be empty.

pub fn new_response(token: Option<u32>) -> Message<'a, RAW, BUFF, COPY, MOVE>

Create a new empty reply. If the request this reply is created for had a token, it should be passed here. The raw data will contain the default value, and all arrays will be empty.

pub fn set_ty(&mut self, ty: MessageTy) -> &mut Self

Sets the message type.

pub fn set_error(&mut self, err: u32) -> &mut Self

Set the error code from a reply.

Panics

Panics if the message is a request.

pub fn error(&self) -> Result<(), Error>

Get the error code from a reply.

Panics

Panics if the message is a request.

pub fn push_raw(&mut self, raw: RAW) -> &mut Self

Sets the raw data of the message.

pub fn raw(&self) -> RAW

Gets the raw data of the message.

pub fn token(&self) -> Option<u32>

Gets the token of a message. This token is used to track IPC call chains. Only present on newer IPC request types.

pub fn push_handle_move(&mut self, handle: Handle) -> &mut Self

Move a handle over IPC. Once the message is sent, the handle will not exist in the current process anymore.

Note

The handle is forgotten as soon as this function is called. If the message is never sent, then the handle will never be closed, causing a handle leak! Furthermore, IPC errors might cause similar problems.

Panics

Panics if attempting to push more handles than there is space for in this message.

pub fn push_handle_copy(&mut self, handle: HandleRef) -> &mut Self

Copy a handle over IPC. The remote process will have a handle that points to the same object.

Panics

Panics if attempting to push more handles than there is space for in this message.

pub fn pop_handle_move(&mut self) -> Result<Handle, Error>

Retrieve a moved handle from this IPC message. Those are popped in the order they were inserted.

Errors

Returns an InvalidMoveHandleCount if attempting to pop more handles than this message has.

pub fn pop_handle_copy(&mut self) -> Result<Handle, Error>

Retrieve a copied handle from this IPC message. Those are popped in the order they were inserted.

Errors

Returns an InvalidCopyHandleCount if attempting to pop more handles than this message has.

pub fn pop_pid(&mut self) -> Result<Pid, Error>

Retrieve the PID of the remote process (if sent at all). This message should only be called once.

Errors

Returns a PidMissing if attempting to pop a Pid from a message that has none, or if attempting to pop a Pid twice.

pub fn pack(self, data: &mut [u8])

Packs this IPC Message to an IPC buffer.

pub fn unpack(data: &[u8]) -> Message<'a, RAW, BUFF, COPY, MOVE>

Parse the passed buffer into an IPC Message.